# (Semi-)vectorize `includes`