# ❓ Draft PR
Not sure if this is good, because it optimizes best cases, whereas pessimizing some worse cases.
Unlike `std::search` vectorization:
 * The set of worse cases seems larger.
 * For `std::search` there's intuitive understanding of an "usual" case: it is "usual" language text search. The "usual" case for `includes` is not known.

I'm not implementing the following for a draft PR, would like the approach to be approved or rejected first:
 * ➕ unsigned types (matter of few adjustments, although this means that unsigned will be less efficient)
 * ✔️ test coverage (the benchmark gives some minimal coverage by checking the expected result for every case)

# ½ Half-vectorization approach

 * Iterate haystack by vector-sized units
 * Skip the whole vector, if all elements are less than the current needle element
 * Skip the vector part is some elements are less than the current needle element
 * Proceed the rest of the current vector in a scalar way

# 😾 Which full vectorization approaches considered and failed

Generally, any of these aproaches result in major slowdown for the following case:
 * Hay: `0,0,1,1,2,2,...999,999`
 * Needle: `0,1,2,...999`

The approaches considered:
* Compare the needle for equality using vector comparison, skip all equal elements. This ends up being close to scalar but using inefficient instructions, when the needle is large and spread nearly equally
* Use `*cmpestr*` in any-to-any match mode. Could possibly have worked if there were no duplicates to consider
* Use `remove`/`unique` approach in a loop. When the needle is large and spread nearly equally, it takes up to half vector size iterations, which is very slow

# ⏱️ Benchmark results

Benchmark                                              |    Before    |   After    | Speedup
-------------------------------------------------------|--------------|------------|----------
bm_includes<int8_t, alg_type::std_fn>/3000/3/0/1       |    370 ns    |   36.8 ns  |   10.05
bm_includes<int8_t, alg_type::std_fn>/3000/22/0/1      |    586 ns    |   60.0 ns  |   9.77
bm_includes<int8_t, alg_type::std_fn>/3000/105/0/1     |    558 ns    |   86.4 ns  |   6.46
bm_includes<int8_t, alg_type::std_fn>/3000/1504/0/1    |    954 ns    |   1026 ns  |   0.93
bm_includes<int8_t, alg_type::std_fn>/3000/2750/0/1    |   1372 ns    |   1488 ns  |   0.92
bm_includes<int8_t, alg_type::std_fn>/300/3/0/1        |   56.3 ns    |   5.55 ns  |   10.14
bm_includes<int8_t, alg_type::std_fn>/300/22/0/1       |   91.6 ns    |   23.9 ns  |   3.83
bm_includes<int8_t, alg_type::std_fn>/300/105/0/1      |    119 ns    |   83.5 ns  |   1.43
bm_includes<int8_t, alg_type::std_fn>/300/290/0/1      |    153 ns    |    231 ns  |   0.66
bm_includes<int8_t, alg_type::std_fn>/3000/3/0/0       |    363 ns    |   36.2 ns  |   10.03
bm_includes<int8_t, alg_type::std_fn>/3000/22/0/0      |    369 ns    |   45.4 ns  |   8.13
bm_includes<int8_t, alg_type::std_fn>/3000/105/0/0     |    399 ns    |   55.1 ns  |   7.24
bm_includes<int8_t, alg_type::std_fn>/3000/1504/0/0    |    560 ns    |    584 ns  |   0.96
bm_includes<int8_t, alg_type::std_fn>/3000/2750/0/0    |    728 ns    |    701 ns  |   1.04
bm_includes<int8_t, alg_type::std_fn>/300/3/0/0        |   51.6 ns    |   5.08 ns  |   10.16
bm_includes<int8_t, alg_type::std_fn>/300/22/0/0       |   80.6 ns    |   16.4 ns  |   4.91
bm_includes<int8_t, alg_type::std_fn>/300/105/0/0      |   82.5 ns    |   35.1 ns  |   2.35
bm_includes<int8_t, alg_type::std_fn>/300/290/0/0      |   83.3 ns    |   88.2 ns  |   0.94
bm_includes<int8_t, alg_type::std_fn>/3000/3/1/1       |    367 ns    |   36.6 ns  |   10.03
bm_includes<int8_t, alg_type::std_fn>/3000/22/1/1      |    456 ns    |   66.6 ns  |   6.85
bm_includes<int8_t, alg_type::std_fn>/3000/105/1/1     |    556 ns    |    163 ns  |   3.41
bm_includes<int8_t, alg_type::std_fn>/3000/1504/1/1    |    935 ns    |   1101 ns  |   0.85
bm_includes<int8_t, alg_type::std_fn>/3000/2750/1/1    |   1355 ns    |   1455 ns  |   0.93
bm_includes<int8_t, alg_type::std_fn>/300/3/1/1        |   55.8 ns    |   13.6 ns  |   4.10
bm_includes<int8_t, alg_type::std_fn>/300/22/1/1       |   99.6 ns    |   34.4 ns  |   2.90
bm_includes<int8_t, alg_type::std_fn>/300/105/1/1      |    151 ns    |    105 ns  |   1.44
bm_includes<int8_t, alg_type::std_fn>/300/290/1/1      |    152 ns    |    227 ns  |   0.67
bm_includes<int8_t, alg_type::std_fn>/3000/3/1/0       |    362 ns    |   33.1 ns  |   10.94
bm_includes<int8_t, alg_type::std_fn>/3000/22/1/0      |    374 ns    |   43.5 ns  |   8.60
bm_includes<int8_t, alg_type::std_fn>/3000/105/1/0     |    729 ns    |   94.7 ns  |   7.70
bm_includes<int8_t, alg_type::std_fn>/3000/1504/1/0    |    560 ns    |    581 ns  |   0.96
bm_includes<int8_t, alg_type::std_fn>/3000/2750/1/0    |    719 ns    |    696 ns  |   1.03
bm_includes<int8_t, alg_type::std_fn>/300/3/1/0        |   51.3 ns    |   5.76 ns  |   8.91
bm_includes<int8_t, alg_type::std_fn>/300/22/1/0       |   67.9 ns    |   19.8 ns  |   3.43
bm_includes<int8_t, alg_type::std_fn>/300/105/1/0      |    105 ns    |   57.7 ns  |   1.82
bm_includes<int8_t, alg_type::std_fn>/300/290/1/0      |   83.5 ns    |   81.2 ns  |   1.03
bm_includes<int8_t, alg_type::std_fn>/3000/3/2/1       |    632 ns    |   59.7 ns  |   10.59
bm_includes<int8_t, alg_type::std_fn>/3000/22/2/1      |    988 ns    |    196 ns  |   5.04
bm_includes<int8_t, alg_type::std_fn>/3000/105/2/1     |   1601 ns    |    851 ns  |   1.88
bm_includes<int8_t, alg_type::std_fn>/3000/1504/2/1    |   2653 ns    |   1964 ns  |   1.35
bm_includes<int8_t, alg_type::std_fn>/3000/2750/2/1    |   2140 ns    |   1612 ns  |   1.33
bm_includes<int8_t, alg_type::std_fn>/300/3/2/1        |    107 ns    |   24.4 ns  |   4.39
bm_includes<int8_t, alg_type::std_fn>/300/22/2/1       |    137 ns    |    140 ns  |   0.98
bm_includes<int8_t, alg_type::std_fn>/300/105/2/1      |    244 ns    |    197 ns  |   1.24
bm_includes<int8_t, alg_type::std_fn>/300/290/2/1      |    235 ns    |    162 ns  |   1.45
bm_includes<int8_t, alg_type::std_fn>/3000/3/2/0       |    378 ns    |   38.1 ns  |   9.92
bm_includes<int8_t, alg_type::std_fn>/3000/22/2/0      |    529 ns    |    123 ns  |   4.30
bm_includes<int8_t, alg_type::std_fn>/3000/105/2/0     |    745 ns    |    420 ns  |   1.77
bm_includes<int8_t, alg_type::std_fn>/3000/1504/2/0    |   1324 ns    |    973 ns  |   1.36
bm_includes<int8_t, alg_type::std_fn>/3000/2750/2/0    |    838 ns    |    801 ns  |   1.05
bm_includes<int8_t, alg_type::std_fn>/300/3/2/0        |   65.7 ns    |   16.2 ns  |   4.06
bm_includes<int8_t, alg_type::std_fn>/300/22/2/0       |   72.5 ns    |   78.6 ns  |   0.92
bm_includes<int8_t, alg_type::std_fn>/300/105/2/0      |    120 ns    |   99.6 ns  |   1.20
bm_includes<int8_t, alg_type::std_fn>/300/290/2/0      |    118 ns    |   83.3 ns  |   1.42
bm_includes<int8_t, alg_type::std_fn>/3000/3/3/1       |    668 ns    |   72.8 ns  |   9.18
bm_includes<int8_t, alg_type::std_fn>/3000/22/3/1      |    976 ns    |    216 ns  |   4.52
bm_includes<int8_t, alg_type::std_fn>/3000/105/3/1     |   1427 ns    |    638 ns  |   2.24
bm_includes<int8_t, alg_type::std_fn>/3000/1504/3/1    |   2476 ns    |   2065 ns  |   1.20
bm_includes<int8_t, alg_type::std_fn>/3000/2750/3/1    |   2336 ns    |   1611 ns  |   1.45
bm_includes<int8_t, alg_type::std_fn>/300/3/3/1        |    117 ns    |   24.4 ns  |   4.80
bm_includes<int8_t, alg_type::std_fn>/300/22/3/1       |    154 ns    |    109 ns  |   1.41
bm_includes<int8_t, alg_type::std_fn>/300/105/3/1      |    228 ns    |    206 ns  |   1.11
bm_includes<int8_t, alg_type::std_fn>/300/290/3/1      |    210 ns    |    162 ns  |   1.30
bm_includes<int8_t, alg_type::std_fn>/3000/3/3/0       |    504 ns    |   56.6 ns  |   8.90
bm_includes<int8_t, alg_type::std_fn>/3000/22/3/0      |    608 ns    |    120 ns  |   5.07
bm_includes<int8_t, alg_type::std_fn>/3000/105/3/0     |    808 ns    |    326 ns  |   2.48
bm_includes<int8_t, alg_type::std_fn>/3000/1504/3/0    |   1229 ns    |    976 ns  |   1.26
bm_includes<int8_t, alg_type::std_fn>/3000/2750/3/0    |    866 ns    |    820 ns  |   1.06
bm_includes<int8_t, alg_type::std_fn>/300/3/3/0        |   87.7 ns    |   17.6 ns  |   4.98
bm_includes<int8_t, alg_type::std_fn>/300/22/3/0       |   71.8 ns    |   62.5 ns  |   1.15
bm_includes<int8_t, alg_type::std_fn>/300/105/3/0      |    117 ns    |    101 ns  |   1.16
bm_includes<int8_t, alg_type::std_fn>/300/290/3/0      |   79.2 ns    |   81.9 ns  |   0.97
bm_includes<int16_t, alg_type::std_fn>/3000/3/0/1      |    373 ns    |   54.4 ns  |   6.86
bm_includes<int16_t, alg_type::std_fn>/3000/22/0/1     |    405 ns    |   64.7 ns  |   6.26
bm_includes<int16_t, alg_type::std_fn>/3000/105/0/1    |    426 ns    |    109 ns  |   3.91
bm_includes<int16_t, alg_type::std_fn>/3000/1504/0/1   |    909 ns    |    788 ns  |   1.15
bm_includes<int16_t, alg_type::std_fn>/3000/2750/0/1   |   1341 ns    |   1378 ns  |   0.97
bm_includes<int16_t, alg_type::std_fn>/300/3/0/1       |   55.8 ns    |   12.1 ns  |   4.61
bm_includes<int16_t, alg_type::std_fn>/300/22/0/1      |   75.3 ns    |   24.3 ns  |   3.10
bm_includes<int16_t, alg_type::std_fn>/300/105/0/1     |    105 ns    |   71.7 ns  |   1.46
bm_includes<int16_t, alg_type::std_fn>/300/290/0/1     |    154 ns    |    162 ns  |   0.95
bm_includes<int16_t, alg_type::std_fn>/3000/3/0/0      |    367 ns    |   52.8 ns  |   6.95
bm_includes<int16_t, alg_type::std_fn>/3000/22/0/0     |    378 ns    |   54.6 ns  |   6.92
bm_includes<int16_t, alg_type::std_fn>/3000/105/0/0    |    395 ns    |   77.2 ns  |   5.12
bm_includes<int16_t, alg_type::std_fn>/3000/1504/0/0   |    558 ns    |    416 ns  |   1.34
bm_includes<int16_t, alg_type::std_fn>/3000/2750/0/0   |    703 ns    |    699 ns  |   1.01
bm_includes<int16_t, alg_type::std_fn>/300/3/0/0       |   52.6 ns    |   10.7 ns  |   4.92
bm_includes<int16_t, alg_type::std_fn>/300/22/0/0      |   64.4 ns    |   18.0 ns  |   3.58
bm_includes<int16_t, alg_type::std_fn>/300/105/0/0     |   81.8 ns    |   42.6 ns  |   1.92
bm_includes<int16_t, alg_type::std_fn>/300/290/0/0     |   86.2 ns    |   83.7 ns  |   1.03
bm_includes<int16_t, alg_type::std_fn>/3000/3/1/1      |    369 ns    |   55.8 ns  |   6.61
bm_includes<int16_t, alg_type::std_fn>/3000/22/1/1     |    400 ns    |   82.1 ns  |   4.87
bm_includes<int16_t, alg_type::std_fn>/3000/105/1/1    |    780 ns    |    190 ns  |   4.11
bm_includes<int16_t, alg_type::std_fn>/3000/1504/1/1   |    909 ns    |    805 ns  |   1.13
bm_includes<int16_t, alg_type::std_fn>/3000/2750/1/1   |   1342 ns    |   1375 ns  |   0.98
bm_includes<int16_t, alg_type::std_fn>/300/3/1/1       |   55.0 ns    |   14.7 ns  |   3.74
bm_includes<int16_t, alg_type::std_fn>/300/22/1/1      |   85.6 ns    |   34.9 ns  |   2.45
bm_includes<int16_t, alg_type::std_fn>/300/105/1/1     |    185 ns    |    111 ns  |   1.67
bm_includes<int16_t, alg_type::std_fn>/300/290/1/1     |    154 ns    |    162 ns  |   0.95
bm_includes<int16_t, alg_type::std_fn>/3000/3/1/0      |    366 ns    |   56.6 ns  |   6.47
bm_includes<int16_t, alg_type::std_fn>/3000/22/1/0     |    380 ns    |   65.1 ns  |   5.84
bm_includes<int16_t, alg_type::std_fn>/3000/105/1/0    |    477 ns    |    123 ns  |   3.88
bm_includes<int16_t, alg_type::std_fn>/3000/1504/1/0   |    558 ns    |    417 ns  |   1.34
bm_includes<int16_t, alg_type::std_fn>/3000/2750/1/0   |    703 ns    |    701 ns  |   1.00
bm_includes<int16_t, alg_type::std_fn>/300/3/1/0       |   51.8 ns    |   12.2 ns  |   4.25
bm_includes<int16_t, alg_type::std_fn>/300/22/1/0      |   65.7 ns    |   22.2 ns  |   2.96
bm_includes<int16_t, alg_type::std_fn>/300/105/1/0     |    109 ns    |   58.0 ns  |   1.88
bm_includes<int16_t, alg_type::std_fn>/300/290/1/0     |   86.5 ns    |   85.4 ns  |   1.01
bm_includes<int16_t, alg_type::std_fn>/3000/3/2/1      |    632 ns    |    104 ns  |   6.08
bm_includes<int16_t, alg_type::std_fn>/3000/22/2/1     |   1017 ns    |    223 ns  |   4.56
bm_includes<int16_t, alg_type::std_fn>/3000/105/2/1    |   1418 ns    |    659 ns  |   2.15
bm_includes<int16_t, alg_type::std_fn>/3000/1504/2/1   |   1792 ns    |   1815 ns  |   0.99
bm_includes<int16_t, alg_type::std_fn>/3000/2750/2/1   |   1842 ns    |   1891 ns  |   0.97
bm_includes<int16_t, alg_type::std_fn>/300/3/2/1       |    106 ns    |   29.3 ns  |   3.62
bm_includes<int16_t, alg_type::std_fn>/300/22/2/1      |    140 ns    |    106 ns  |   1.32
bm_includes<int16_t, alg_type::std_fn>/300/105/2/1     |    220 ns    |    209 ns  |   1.05
bm_includes<int16_t, alg_type::std_fn>/300/290/2/1     |    254 ns    |    170 ns  |   1.49
bm_includes<int16_t, alg_type::std_fn>/3000/3/2/0      |    382 ns    |   67.7 ns  |   5.64
bm_includes<int16_t, alg_type::std_fn>/3000/22/2/0     |    536 ns    |    110 ns  |   4.87
bm_includes<int16_t, alg_type::std_fn>/3000/105/2/0    |    737 ns    |    324 ns  |   2.27
bm_includes<int16_t, alg_type::std_fn>/3000/1504/2/0   |    993 ns    |    924 ns  |   1.07
bm_includes<int16_t, alg_type::std_fn>/3000/2750/2/0   |    846 ns    |    923 ns  |   0.92
bm_includes<int16_t, alg_type::std_fn>/300/3/2/0       |   67.0 ns    |   18.6 ns  |   3.60
bm_includes<int16_t, alg_type::std_fn>/300/22/2/0      |   91.8 ns    |   60.0 ns  |   1.53
bm_includes<int16_t, alg_type::std_fn>/300/105/2/0     |    121 ns    |   98.0 ns  |   1.23
bm_includes<int16_t, alg_type::std_fn>/300/290/2/0     |    129 ns    |   86.8 ns  |   1.49
bm_includes<int16_t, alg_type::std_fn>/3000/3/3/1      |    665 ns    |    114 ns  |   5.83
bm_includes<int16_t, alg_type::std_fn>/3000/22/3/1     |    980 ns    |    203 ns  |   4.83
bm_includes<int16_t, alg_type::std_fn>/3000/105/3/1    |   1423 ns    |    542 ns  |   2.63
bm_includes<int16_t, alg_type::std_fn>/3000/1504/3/1   |   2795 ns    |   2256 ns  |   1.24
bm_includes<int16_t, alg_type::std_fn>/3000/2750/3/1   |   2173 ns    |   1685 ns  |   1.29
bm_includes<int16_t, alg_type::std_fn>/300/3/3/1       |    118 ns    |   29.3 ns  |   4.03
bm_includes<int16_t, alg_type::std_fn>/300/22/3/1      |    144 ns    |   97.3 ns  |   1.48
bm_includes<int16_t, alg_type::std_fn>/300/105/3/1     |    230 ns    |    200 ns  |   1.15
bm_includes<int16_t, alg_type::std_fn>/300/290/3/1     |    216 ns    |    174 ns  |   1.24
bm_includes<int16_t, alg_type::std_fn>/3000/3/3/0      |    504 ns    |   82.4 ns  |   6.12
bm_includes<int16_t, alg_type::std_fn>/3000/22/3/0     |    608 ns    |    116 ns  |   5.24
bm_includes<int16_t, alg_type::std_fn>/3000/105/3/0    |    767 ns    |    273 ns  |   2.81
bm_includes<int16_t, alg_type::std_fn>/3000/1504/3/0   |   1393 ns    |   1103 ns  |   1.26
bm_includes<int16_t, alg_type::std_fn>/3000/2750/3/0   |   1085 ns    |    855 ns  |   1.27
bm_includes<int16_t, alg_type::std_fn>/300/3/3/0       |   88.9 ns    |   22.3 ns  |   3.99
bm_includes<int16_t, alg_type::std_fn>/300/22/3/0      |   69.3 ns    |   52.5 ns  |   1.32
bm_includes<int16_t, alg_type::std_fn>/300/105/3/0     |    119 ns    |   93.6 ns  |   1.27
bm_includes<int16_t, alg_type::std_fn>/300/290/3/0     |    111 ns    |   89.7 ns  |   1.24
bm_includes<int32_t, alg_type::std_fn>/3000/3/0/1      |    370 ns    |   89.2 ns  |   4.15
bm_includes<int32_t, alg_type::std_fn>/3000/22/0/1     |    392 ns    |   98.6 ns  |   3.98
bm_includes<int32_t, alg_type::std_fn>/3000/105/0/1    |    424 ns    |    153 ns  |   2.77
bm_includes<int32_t, alg_type::std_fn>/3000/1504/0/1   |    959 ns    |    861 ns  |   1.11
bm_includes<int32_t, alg_type::std_fn>/3000/2750/0/1   |   1351 ns    |   1491 ns  |   0.91
bm_includes<int32_t, alg_type::std_fn>/300/3/0/1       |   55.5 ns    |   16.6 ns  |   3.34
bm_includes<int32_t, alg_type::std_fn>/300/22/0/1      |   74.9 ns    |   26.2 ns  |   2.86
bm_includes<int32_t, alg_type::std_fn>/300/105/0/1     |    104 ns    |   74.7 ns  |   1.39
bm_includes<int32_t, alg_type::std_fn>/300/290/0/1     |    152 ns    |    169 ns  |   0.90
bm_includes<int32_t, alg_type::std_fn>/3000/3/0/0      |    367 ns    |   88.0 ns  |   4.17
bm_includes<int32_t, alg_type::std_fn>/3000/22/0/0     |    451 ns    |   92.6 ns  |   4.87
bm_includes<int32_t, alg_type::std_fn>/3000/105/0/0    |    392 ns    |    126 ns  |   3.11
bm_includes<int32_t, alg_type::std_fn>/3000/1504/0/0   |    557 ns    |    458 ns  |   1.22
bm_includes<int32_t, alg_type::std_fn>/3000/2750/0/0   |    704 ns    |    750 ns  |   0.94
bm_includes<int32_t, alg_type::std_fn>/300/3/0/0       |   51.2 ns    |   15.3 ns  |   3.35
bm_includes<int32_t, alg_type::std_fn>/300/22/0/0      |   67.6 ns    |   19.9 ns  |   3.40
bm_includes<int32_t, alg_type::std_fn>/300/105/0/0     |   78.5 ns    |   39.1 ns  |   2.01
bm_includes<int32_t, alg_type::std_fn>/300/290/0/0     |   82.5 ns    |   91.3 ns  |   0.90
bm_includes<int32_t, alg_type::std_fn>/3000/3/1/1      |    370 ns    |   92.1 ns  |   4.02
bm_includes<int32_t, alg_type::std_fn>/3000/22/1/1     |    408 ns    |    113 ns  |   3.61
bm_includes<int32_t, alg_type::std_fn>/3000/105/1/1    |    576 ns    |    204 ns  |   2.82
bm_includes<int32_t, alg_type::std_fn>/3000/1504/1/1   |    909 ns    |    855 ns  |   1.06
bm_includes<int32_t, alg_type::std_fn>/3000/2750/1/1   |   1344 ns    |   1490 ns  |   0.90
bm_includes<int32_t, alg_type::std_fn>/300/3/1/1       |   55.6 ns    |   18.3 ns  |   3.04
bm_includes<int32_t, alg_type::std_fn>/300/22/1/1      |   90.4 ns    |   34.9 ns  |   2.59
bm_includes<int32_t, alg_type::std_fn>/300/105/1/1     |    155 ns    |    105 ns  |   1.48
bm_includes<int32_t, alg_type::std_fn>/300/290/1/1     |    151 ns    |    169 ns  |   0.89
bm_includes<int32_t, alg_type::std_fn>/3000/3/1/0      |    367 ns    |   90.6 ns  |   4.05
bm_includes<int32_t, alg_type::std_fn>/3000/22/1/0     |    511 ns    |   96.5 ns  |   5.30
bm_includes<int32_t, alg_type::std_fn>/3000/105/1/0    |    434 ns    |    153 ns  |   2.84
bm_includes<int32_t, alg_type::std_fn>/3000/1504/1/0   |    556 ns    |    463 ns  |   1.20
bm_includes<int32_t, alg_type::std_fn>/3000/2750/1/0   |    700 ns    |    749 ns  |   0.93
bm_includes<int32_t, alg_type::std_fn>/300/3/1/0       |   51.8 ns    |   17.8 ns  |   2.91
bm_includes<int32_t, alg_type::std_fn>/300/22/1/0      |   72.8 ns    |   26.2 ns  |   2.78
bm_includes<int32_t, alg_type::std_fn>/300/105/1/0     |    103 ns    |   58.0 ns  |   1.78
bm_includes<int32_t, alg_type::std_fn>/300/290/1/0     |   82.5 ns    |   92.7 ns  |   0.89
bm_includes<int32_t, alg_type::std_fn>/3000/3/2/1      |    635 ns    |    168 ns  |   3.78
bm_includes<int32_t, alg_type::std_fn>/3000/22/2/1     |   1003 ns    |    246 ns  |   4.08
bm_includes<int32_t, alg_type::std_fn>/3000/105/2/1    |   1392 ns    |    545 ns  |   2.55
bm_includes<int32_t, alg_type::std_fn>/3000/1504/2/1   |   1185 ns    |   1867 ns  |   0.63
bm_includes<int32_t, alg_type::std_fn>/3000/2750/2/1   |   1495 ns    |   1643 ns  |   0.91
bm_includes<int32_t, alg_type::std_fn>/300/3/2/1       |    107 ns    |   28.8 ns  |   3.72
bm_includes<int32_t, alg_type::std_fn>/300/22/2/1      |    140 ns    |   97.8 ns  |   1.43
bm_includes<int32_t, alg_type::std_fn>/300/105/2/1     |    222 ns    |    194 ns  |   1.14
bm_includes<int32_t, alg_type::std_fn>/300/290/2/1     |    236 ns    |    181 ns  |   1.30
bm_includes<int32_t, alg_type::std_fn>/3000/3/2/0      |    379 ns    |    105 ns  |   3.61
bm_includes<int32_t, alg_type::std_fn>/3000/22/2/0     |    524 ns    |    122 ns  |   4.30
bm_includes<int32_t, alg_type::std_fn>/3000/105/2/0    |    683 ns    |    282 ns  |   2.42
bm_includes<int32_t, alg_type::std_fn>/3000/1504/2/0   |    594 ns    |    935 ns  |   0.64
bm_includes<int32_t, alg_type::std_fn>/3000/2750/2/0   |    689 ns    |    830 ns  |   0.83
…