@AlexGuteniev 2025-06-29
> @AlexGuteniev After looking at your benchmark results, I'm fine with this approach. We always prefer pure wins but in this case I think the relatively small and infrequent regressions are fine given the improved cases.

Thanks for confirmation, added the missing parts, now ready for review.